% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogo_disponible.R
\name{catalogo_disponible}
\alias{catalogo_disponible}
\title{Muestra las series disponibles en el portal de datos abiertos de la Ciudad de México}
\usage{
catalogo_disponible()
}
\description{
Muestra las series disponibles en el portal de datos abiertos de la Ciudad de México
}
\examples{
catalogo_disponible()
}
