% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Elija un ID de las series disponibles y descargue la base de datos, el diccionario de datos y los metadatos}
\usage{
descargar_datos(id)
}
\arguments{
\item{id}{numeric variable}
}
\description{
Elija un ID de las series disponibles y descargue la base de datos, el diccionario de datos y los metadatos
}
\examples{
# id <- 3
# descargar_datos(id = id)
# Puede realizar descargas masivas creando un vector con
# los ids de las series que desea descargar y un ciclo for:
# Ejemplo 1
# ids <- 105:110
# for(i in ids) {descargar_datos(i)}
# Ejemplo 2
# series <- c(2, 94, 170, 241)
# for(i in series) {descargar_datos(i)}
}
