% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_metadatos.R
\name{obtener_metadatos}
\alias{obtener_metadatos}
\title{Elija un ID de las series disponibles y cargue los metadatos de la serie seleccionada al enviroment de R como un data.frame}
\usage{
obtener_metadatos(id)
}
\arguments{
\item{id}{numeric variable}
}
\description{
Elija un ID de las series disponibles y cargue los metadatos de la serie seleccionada al enviroment de R como un data.frame
}
\examples{
id <- 3
metadatos <- obtener_metadatos(id = id)
metadatos
}
