% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Elija un ID de las series disponibles y descargue la base de datos, el diccionario de datos y los metadatos}
\usage{
descargar_datos(id)
}
\arguments{
\item{id}{numeric variable}
}
\description{
Elija un ID de las series disponibles y descargue la base de datos, el diccionario de datos y los metadatos
}
\examples{
# id <- 3
# descargar_datos(id = id)
}
